Initial files
Added the modelNN file which goes through the initial neural net testing.  The utils module, which has some data prep functions.  The environment.yaml, which ensures we are operating from the same environment (note - that will need to be updated with new packages and the environment I am presently using).